requirements/common.txt
#common requirements
django-bootstrap-toolkit
django-extensions
sentry==dev
django-debug-toolbar
